+++ web/src/pages/agent/form/agent-form/tool-popover/tool-command.tsx
@@ -42,6 +42,7 @@ const Menus = [
       Operator.Invoke,
       Operator.Crawler,
       Operator.Code,
+      Operator.Retrieval,
     ],
   },
 ];
