+++ web/src/pages/agent/form/tool-form/constant.ts
@@ -12,9 +12,9 @@ import GithubForm from '../github-form';
 import GoogleForm from '../google-form';
 import GoogleScholarForm from '../google-scholar-form';
 import PubMedForm from '../pubmed-form';
-import RetrievalForm from '../retrieval-form/next';
 import WikipediaForm from '../wikipedia-form';
 import YahooFinanceForm from '../yahoo-finance-form';
+import RetrievalForm from './retrieval-form';
 import TavilyForm from './tavily-form';
 
 export const ToolFormConfigMap = {
